requirements.txt
pandas
nltk
scikit-learn
pillow
streamlit
torch 
torchvision
-e git+https://github.com/pytorch/fairseq.git#egg=fairseq